requirements.txt
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
Pillow==9.5.0
requests==2.31.0
scikit-learn==1.3.0
streamlit==1.25.0
tensorflow==2.13.0
tqdm==4.66.1
 